requirements.txt
opencv-python-headless>=4.10.0.84
ultralytics>=8.3.0 # For YOLOv8
numpy>=1.23
torch>=2.4.1 # Adjust based on CUDA/CPU needs; see https://pytorch.org/get-started/locally/
torchvision>=0.19.1
redis>=5.0.8
numpy==1.26.4
pyyaml==6.0.2
scipy>=1.9
filterpy>=1.4.5
